feat(tests): add migration tests
